[oscillators] get_temp consistency

+++ src/oscillators/sim_oscillator.c
@@ -80,12 +80,11 @@ static int sim_oscillator_save(struct oscillator *oscillator)
 	return -ENOSYS;
 }
 
-static int sim_oscillator_get_temp(struct oscillator *oscillator,
-		uint16_t *temp)
+static int sim_oscillator_get_temp(struct oscillator *oscillator, double *temp)
 {
 	*temp = (rand() % (55 - 10)) + 10;
 
-	log_info("%s(%p, %u)", __func__, oscillator, *temp);
+	log_info("%s(%p, %g)", __func__, oscillator, *temp);
 
 	return 0;
 }
